+++ src/main/java/co/ratethem/payload/ReviewAddRequest.java
@@ -15,25 +15,25 @@ public class ReviewAddRequest {
 
     //HR Start
     private String hrName;
-    private double hrIceBrake;
-    private double hrAttitude;
-    private double hrPunctuality;
-    private double hrImpression;
+    private int hrIceBrake;
+    private int hrAttitude;
+    private int hrPunctuality;
+    private int hrImpression;
     private String hrComment;
     //HR End
 
     //Tech Start
     private String techInterviewerName;
-    private double techIceBrake;
-    private double techAttitude;
-    private double techQuestionsQuality;
-    private double techImpression;
+    private int techIceBrake;
+    private int techAttitude;
+    private int techQuestionsQuality;
+    private int techImpression;
     private String techComment;
     //Tech End
 
     //Feedback Start
-    private double feedbackOnTime;
-    private double feedbackDetailization;
+    private int feedbackOnTime;
+    private int feedbackDetailization;
     private String feedbackComment;
     //Feedback End
 }
